refactor: rename user customError or errors
