+++ src/pages/communities/[slug]/challenges/[challenge_id]/submissions/[submission_id].tsx
@@ -15,7 +15,7 @@ import { Challenge } from "@/types/course";
 import { initChallengeNavigationMenu } from "@/store/feature/communities/navigation.slice";
 import useNavigation from "@/hooks/useNavigation";
 import { fetchChallenge } from "@/store/services/communities/challenges";
-import { NotFoundError } from "@/utilities/customErrors/NotFoundError";
+import { NotFoundError } from "@/utilities/errors/NotFoundError";
 
 export default function SubmissionPage(props: { pageProps: { challenge: Challenge } }) {
   const dispatch = useDispatch();
